// async-race/src/utils/enums.ts
export const enum PageIds {
  GaragePage = 'garage',
  WinnersPage = 'winners',
  DefaultId = 'current-page',
}

export const enum Components {
  Header = 'header',
  Main = 'main',
  Div = 'div',
  Section = 'section',
  Button = 'button',
  Table = 'table',
  Row = 'tr',
  Footer = 'footer',
}

export const enum ClassNames {
  Container = 'container',
  MainContainer = 'main__container',
  PageButtons = 'page__buttons',
  CarImg = 'car__img',
  Flag = 'flag',
  Message = 'message',
  Hidden = '_hidden',
}

export const enum ButtonsIds {
  Create = 'create',
  Update = 'update',
  Race = 'race',
  Reset = 'reset',
  Generate = 'generate',
  PrevGarage = 'prev-garage',
  NextGarage = 'next-garage',
  PrevWinners = 'prev-winners',
  NextWinners = 'next-winners',
  Wins = 'wins',
  Time = 'time',
}

export const enum ButtonsClasses {
  Select = 'select',
  Remove = 'remove',
  Start = 'start',
  Stop = 'stop',
  ButtonDisabled = 'button_disabled',
  PageDisabled = 'page_disabled',
  Disabled = '_disabled',
  Sort = 'sort',
  ASC = 'ASC',
  DESC = 'DESC',
  Page = 'button_page',
}

export const enum InputsIds {
  Name = 'name',
  Color = 'color',
  NewName = 'new-name',
  NewColor = 'new-color',
}

export const enum DefaultValues {
  Name = '',
  Color = '#ffffff',
  CarsLimit = 7,
  WinnersLimit = 10,
  Sort = 'id',
  Order = 'ASC',
}